fix : 비공개 레이아웃으로 헤더 요소 이동 (#2)
