+++ src/components/Layout/Layout.tsx
@@ -1,23 +1,9 @@
-import { Outlet, ScrollRestoration, useLocation } from 'react-router-dom';
-
-import HeaderHome from '@components/Layout//HeaderHome';
-import HeaderBack from '@components/Layout/HeaderBack';
+import { Outlet, ScrollRestoration } from 'react-router-dom';
 
 const Layout = () => {
-  const location = useLocation();
-  const pathname = location.pathname;
-  const pathParts = pathname.split('/').slice(1);
-
   return (
     <>
       <div className="h-[100vh] min-w-[373px] bg-yellow-50">
-        {HEADERS.filter((header) => header.pathname === pathParts[0]).map(
-          (component) => (
-            <header key={component.pathname} className="h-[60px] bg-amber-300">
-              {component.component}
-            </header>
-          ),
-        )}
         <main className="h-full">
           <Outlet />
         </main>
@@ -28,9 +14,3 @@ const Layout = () => {
 };
 
 export default Layout;
-
-const HEADERS = [
-  { pathname: 'home', component: <HeaderHome /> },
-  { pathname: 'question', component: <HeaderBack /> },
-  { pathname: 'vote', component: <HeaderBack /> },
-];
